requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
channels==3.0.4
configparser==5.2.0
dj-database-url==0.5.0
Django==4.0.1
django-heroku==0.3.1
django-jazzmin==2.4.8
djangorestframework==3.13.1
django-rest-swagger==2.2.0
drf-yasg==1.20.0
gunicorn==20.1.0
psycopg2==2.9.3
pandas==1.4.2
python-decouple==3.5
selenium==4.1.0
sqlparse==0.4.2
tzdata==2021.5
virtualenv==20.10.0
webdriver-manager==3.5.4
whitenoise==5.3.0
